pyproject.toml
[tool.poetry]
name = "crm"
version = "0.1.0"
description = ""
authors = ["Momotov Kirill"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
djangorestframework = "^3.15.1"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
